% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvgz.r
\name{csv.gz}
\alias{csv.gz}
\title{CSV to GZ Function}
\usage{
csv.gz(submission, filename = "submission")
}
\arguments{
\item{submission}{data.frame or data.table data type}

\item{filename}{file name you want to save it as - string data type}
}
\description{
This function allows you to save a csv file in gz format.
}
\examples{
csv.gz
}
\keyword{csv}
