% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb.importance.table.R
\name{xgb.importance.table}
\alias{xgb.importance.table}
\title{XGB Importance Function}
\usage{
xgb.importance.table(model, feature_names = NULL, gainColor = "lightgreen",
  digits = 6, coverColor = "lightblue", freqColor = "lightgrey",
  pageLength = 20, lengthMenu = c(5, 10, 15, 20, 25, 50, 100, 250, 500))
}
\arguments{
\item{gainColor}{Type: character. Defines the display color for Gain column.}

\item{digits}{Type: integer. The length of decimal place to display results.}

\item{coverColor}{Type: character. Defines the display color of cover Column.}

\item{freqColor}{Type: character. Defines the display color for frequency column.}

\item{pageLength}{Type: integer. Defines the number of rows per page.}

\item{Model}{Type: XGBoost Model. Must be defined.}

\item{features}{Type: character. Vector. Must be defined.}

\item{menuLength}{Type: integer. Changes the option in the page length menu.}
}
\description{
This function allows you to graph the gain, cover and frequency of an XGB model.
}
\examples{
cat_function()
}
\keyword{xgb.importance}
